Lots of WIP stuff
Repo is currently a mess!
- Some work on dynamic trees
- Redoing the way queries and casts work to make history rollback more
memory efficient
