+++ Unity/Assets/Scripts/Demo/GizmoDrawer.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+
+using UnityEngine;
+
+public class GizmoDrawer : MonoBehaviour 
+{
+  public static GizmoDrawer Instance;
+
+  public Vector2 start;
+  public Vector2 end;
+
+  void Awake()
+  {
+    Instance = this;
+  }
+
+  void OnDrawGizmos()
+  {
+    Gizmos.color = Color.magenta;
+    Gizmos.DrawLine(start, end);
+  }
+}
