--- Unity/Assets/VolatilePhysics.dll.meta
@@ -1,20 +1,0 @@
-fileFormatVersion: 2
-guid: 21983187b3e9e1d4cbfc237873d5108b
-timeCreated: 1453052612
-licenseType: Free
-PluginImporter:
-  serializedVersion: 1
-  iconMap: {}
-  executionOrder: {}
-  isPreloaded: 0
-  platformData:
-    Any:
-      enabled: 1
-      settings: {}
-    Editor:
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
